+++ index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="index.css">
     <title>Awesome Books</title>
 </head>
 
@@ -71,7 +71,7 @@
     </div>
    </div>
 </footer>
-    <script src="main.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 
 </html>
